resolve error issue as regards to deleting most of the file, everything is good except the header shadow
