--- src/screens/signup/phoneVerification/style.ts
@@ -1,60 +1,0 @@
-import styled from 'styled-components/native';
-import { Dimensions } from 'react-native';
-import { theme } from '../../../theme/types';
-
-const { width } = Dimensions.get('window');
-
-export const Container = styled.View`
-  flex: 1;
-  width: ${width}px;
-  padding: 15px;
-`;
-
-export const Navigate = styled.View`
-  margin-top: 24.5px;
-  width: 100%;
-  flex-direction: row;
-  align-content: center;
-  justify-content: flex-start;
-  height: 18px;
-`;
-
-export const GoBack = styled.TouchableOpacity`
-  height: 18px;
-  width: 18px;
-`;
-
-export const Text = styled.Text`
-  font-size: ${({ theme }) => theme.fonts.MEDIUM_SIZE}px;
-  font-family: ${({ theme }) => theme.fonts.JOST_BOOK};
-  text-align: center;
-  margin-left: 40%;
-`;
-
-export const ButtonStyle = {
-  mainButton: {
-    width: '100%'
-  },
-  textStyle: {
-    fontFamily: theme.fonts.JOST_MEDIUM,
-    fontSize: 15
-  }
-};
-
-export const Header = styled.Text`
-  font-family: ${({ theme }) => theme.fonts.JOST_BOOK};
-  width: 311px;
-  font-size: 30px;
-  text-align: left;
-  height: 43px;
-  margin-top: 170px;
-`;
-
-export const Subtitle = styled.Text`
-  font-size: 20px;
-  font-family: ${({ theme }) => theme.fonts.CORMORANT_MEDIUM};
-  text-align: left;
-  width: 327px;
-  opacity: 0.7;
-  height: 53px;
-`;
