--- src/screens/signup/styles.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components/native';
-
-export const Container = styled.View`
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-  background-color: ${({ theme }) => theme.colors.BG_WHITE_COLOR};
-`;
-
-export const Welcome = styled.Text`
-  font-size: ${({ theme }) => theme.fonts.LARGE_SIZE}px;
-  font-family: ${({ theme }) => theme.fonts.JOST_MEDIUM};
-  color: ${({ theme }) => theme.colors.FONT_DARK_COLOR};
-  text-transform: capitalize;
-`;
